+++ src/main/java/thedarkcolour/exdeorum/compat/PreferredOres.java
@@ -39,7 +39,7 @@ import java.util.Comparator;
 import java.util.Map;
 
 public class PreferredOres {
-    private static final Map<TagKey<Item>, Item> PREFERRED_ORE_ITEMS = new Object2ObjectOpenHashMap<>(11, Hash.DEFAULT_LOAD_FACTOR);
+    private static final Map<TagKey<Item>, Item> PREFERRED_ORE_ITEMS = new Object2ObjectOpenHashMap<>(15, Hash.DEFAULT_LOAD_FACTOR);
 
     static {
         putPreferredOre(EItemTags.ORES_ALUMINUM, EConfig.COMMON.preferredAluminumOre, getDefaultAluminumOre());
@@ -104,7 +104,11 @@ public class PreferredOres {
     }
 
     public static Item getDefaultAluminumOre() {
-        return defaultItem("aluminum_ore", ModIds.ALL_THE_ORES, ModIds.IMMERSIVE_ENGINEERING);
+        var item = defaultItem("aluminum_ore", ModIds.ALL_THE_ORES, ModIds.IMMERSIVE_ENGINEERING, ModIds.NATURE_ARISE);
+        if (item == Items.AIR && ModList.get().isLoaded(ModIds.MODERN_INDUSTRIALIZATION)) {
+            item = defaultItem("bauxite_ore", ModIds.MODERN_INDUSTRIALIZATION);
+        }
+        return item;
     }
 
     public static Item getDefaultCobaltOre() {
@@ -116,15 +120,15 @@ public class PreferredOres {
     }
 
     public static Item getDefaultLeadOre() {
-        return defaultItem("lead_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.MEKANISM, ModIds.RAILCRAFT, ModIds.FACTORIUM, ModIds.NUCLEARCRAFT_NEOTERIC, ModIds.IMMERSIVE_ENGINEERING);
+        return defaultItem("lead_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.MODERN_INDUSTRIALIZATION, ModIds.MEKANISM, ModIds.RAILCRAFT, ModIds.FACTORIUM, ModIds.NUCLEARCRAFT_NEOTERIC, ModIds.IMMERSIVE_ENGINEERING);
     }
 
     public static Item getDefaultPlatinumOre() {
-        return defaultItem("platinum_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.FACTORIUM, ModIds.NUCLEARCRAFT_NEOTERIC);
+        return defaultItem("platinum_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.MODERN_INDUSTRIALIZATION, ModIds.FACTORIUM, ModIds.NUCLEARCRAFT_NEOTERIC);
     }
 
     public static Item getDefaultNickelOre() {
-        var item = defaultItem("nickel_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.THERMAL, ModIds.RAILCRAFT, ModIds.FACTORIUM, ModIds.IMMERSIVE_ENGINEERING);
+        var item = defaultItem("nickel_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.MODERN_INDUSTRIALIZATION, ModIds.THERMAL, ModIds.RAILCRAFT, ModIds.FACTORIUM, ModIds.IMMERSIVE_ENGINEERING);
 
         if (item == Items.AIR && ModList.get().isLoaded(ModIds.IMMERSIVE_ENGINEERING)) {
             item = defaultItem("ore_nickel", ModIds.IMMERSIVE_ENGINEERING);
@@ -134,7 +138,7 @@ public class PreferredOres {
     }
 
     public static Item getDefaultUraniumOre() {
-        var item = defaultItem("uranium_ore", ModIds.ALL_THE_ORES, ModIds.MEKANISM, ModIds.BIGGER_REACTORS, ModIds.NUCLEARCRAFT_NEOTERIC, ModIds.IMMERSIVE_ENGINEERING);
+        var item = defaultItem("uranium_ore", ModIds.ALL_THE_ORES, ModIds.MODERN_INDUSTRIALIZATION, ModIds.MEKANISM, ModIds.BIGGER_REACTORS, ModIds.NUCLEARCRAFT_NEOTERIC, ModIds.IMMERSIVE_ENGINEERING);
 
         if (item == Items.AIR && ModList.get().isLoaded(ModIds.EXTREME_REACTORS)) {
             item = defaultItem("yellorite_ore", ModIds.EXTREME_REACTORS);
@@ -148,7 +152,7 @@ public class PreferredOres {
     }
 
     public static Item getDefaultTinOre() {
-        return defaultItem("tin_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.THERMAL, ModIds.MEKANISM, ModIds.RAILCRAFT, ModIds.FACTORIUM, ModIds.NUCLEARCRAFT_NEOTERIC);
+        return defaultItem("tin_ore", ModIds.ALL_THE_ORES, ModIds.GREG, ModIds.MODERN_INDUSTRIALIZATION, ModIds.THERMAL, ModIds.MEKANISM, ModIds.FORESTRY, ModIds.RAILCRAFT, ModIds.FACTORIUM, ModIds.NUCLEARCRAFT_NEOTERIC);
     }
 
     public static Item getDefaultZincOre() {
@@ -156,7 +160,7 @@ public class PreferredOres {
     }
 
     public static Item getDefaultIridiumOre() {
-        return defaultItem("iridium_ore", ModIds.ALL_THE_ORES);
+        return defaultItem("iridium_ore", ModIds.ALL_THE_ORES, ModIds.MODERN_INDUSTRIALIZATION);
     }
 
     public static Item getDefaultThoriumOre() {