+++ changelog.md
@@ -1,5 +1,6 @@
 ## Ex Deorum 1.45
 - Allow changing drops for Random Armor Trim and Pottery Sherd, also add Tide to possible trims (#138)
+- Add compatibility with Modern Industrialization, Forestry, Nature Arise ()
 
 ## Ex Deorum 1.44
 - Fix Silkworms not applying to certain modded leaves like TFC